+++ chain_harvester/constants.py
@@ -19,6 +19,7 @@ class Network(IntEnum):
     Filecoin = 314
     Mode = 34443
     Scroll = 534352
+    Linea = 59144
     Rari = 1380012617
 
 
@@ -48,6 +49,7 @@ MULTICALL3_ADDRESSES = {
     Network.Filecoin: "0xcA11bde05977b3631167028862bE2a173976CA11",
     Network.Mode: "0xcA11bde05977b3631167028862bE2a173976CA11",
     Network.Scroll: "0xcA11bde05977b3631167028862bE2a173976CA11",
+    Network.Linea: "0xcA11bde05977b3631167028862bE2a173976CA11",
     Network.Rari: "0x4c753F58Ee9E83B38170abAbBEa8B47976C7ee1b",
 }
 
@@ -86,6 +88,9 @@ CHAINS = {
     },
     "scroll": {
         "mainnet": Network.Scroll,
+    },
+    "linea": {
+        "mainnet": Network.Linea,
     },
     "rari": {
         "mainnet": Network.Rari,